Splitting out a few of the components into smaller files
