+++ src/index.js
@@ -1,8 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { v4 as uuidv4 } from "uuid";
+
+// import progressBar from './components/pb';
+import Equipment from './components/equipment';
+import Role from './components/roleSelector'
 
 import "./scss/index.scss";
+
 
 function PageHeading() {
   return (
@@ -45,345 +49,6 @@ class ProgressBar extends React.Component {
     );
   }
 }
-class Role extends React.Component {
-  render() {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="card-body">
-            <h4 className="card-title col">Select your class</h4>
-            <div
-              className="btn-group card-text col"
-              role="group"
-              aria-label="Role"
-            >
-              <button
-                type="button"
-                className="btn btn-melee"
-                onClick={this.props.roleChange}
-                value="melee"
-              >
-                Melee
-              </button>
-              <button
-                type="button"
-                className="btn btn-ranged"
-                onClick={this.props.roleChange}
-                value="ranged"
-              >
-                Ranged
-              </button>
-              <button
-                type="button"
-                className="btn btn-magic"
-                onClick={this.props.roleChange}
-                value="magic"
-              >
-                Magic
-              </button>
-              <button
-                type="button"
-                className="btn btn-summoner"
-                onClick={this.props.roleChange}
-                value="summoner"
-              >
-                Summoner
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Equipment extends React.Component {
-  render() {
-    return (
-      <div className="equipment" id="equipment">
-        <div className="row">
-          <div className="col">
-            <Weapons
-              currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex}
-            />
-          </div>
-          <div className="col">
-            <Armor currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex} />
-          </div>
-        </div>
-        <div className="row mt-4">
-          <div className="col">
-            <Accessories currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex} />
-          </div>
-          <div className="col">
-            <Buffs currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex} />
-          </div>
-        </div>
-        <div className="row mt-4">
-          <div className="col">
-            <Mounts currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex} />
-          </div>
-          <div className="col">
-            <Lights currentRole={this.props.currentRole}
-              gameStageIndex={this.props.gameStageIndex} />
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-// Check this our for some ideas to avoid duplicating code
-// https://stackoverflow.com/questions/36983158/rendering-json-data-in-multiple-react-js-components
-
-function getWeapons(currentRole, gameStageIndex) {
-  let weaponList = [];
-  const weaponData = require("./data/items.json");
-  for (var i = 0; i < weaponData.length; i++) {
-    const weapon = weaponData[i];
-    if (
-      (weapon.role === currentRole || weapon.role === "mixed") &&
-      weapon.gameStageAvailable === gameStageIndex &&
-      weapon.category === "weapons"
-    ) {
-      weaponList.push(weapon);
-    }
-  }
-
-  return weaponList;
-}
-
-function populateWeapons(currentRole, gameStageIndex) {
-  const weaponList = getWeapons(currentRole, gameStageIndex);
-
-  const listItems = currentRole ? (
-    weaponList.map((weapon) => (
-      <tr key={uuidv4()}>
-        <td className="image-cell">
-          <img
-            key={uuidv4()}
-            src={weapon.imgPath}
-            alt={weapon.name}
-            className="mr-3 mt-1"
-            decoding="async"
-          ></img>
-        </td>
-        <td key={uuidv4()}>
-          <a href={weapon.url} key={uuidv4()} target="_blank" rel="noreferrer">
-            {weapon.name}
-          </a>
-        </td>
-      </tr>
-    ))
-  ) : (
-    <tr>
-      <th key="1">Select a class</th>
-    </tr>
-  );
-  return listItems;
-}
-
-function getArmor(currentRole, gameStageIndex) {
-  let armorList = [];
-  const armorData = require("./data/items.json");
-  for (var i = 0; i < armorData.length; i++) {
-    const armor = armorData[i];
-    if (
-      (armor.role === currentRole || armor.role === "mixed") &&
-      armor.gameStageAvailable === gameStageIndex && armor.category === "armor"
-      ) {
-        armorList.push(armor);
-      }
-    }
-    return armorList;
-}
-
-function populateArmor(currentRole, gameStageIndex) {
-  const armorList = getArmor(currentRole, gameStageIndex);
-
-  const listItems = currentRole ? (
-    armorList.map((armor) => (
-      <tr key={uuidv4()}>
-        <td key={uuidv4()}>
-          <img
-            key={uuidv4()}
-            src={armor.imgPath}
-            alt={armor.name}
-            className="mr-3 mt-1"
-            decoding="async"
-          ></img>
-        </td>
-        <td key={uuidv4()}>
-          <a href={armor.url} key={uuidv4()} target="_blank" rel="noreferrer">
-            {armor.name}
-          </a>
-        </td>
-      </tr>
-    ))
-  ) : (
-    <tr>
-      <th key="1">Select a class</th>
-    </tr>
-  );
-  
-  return listItems;
-}
-
-class Weapons extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Weapons
-        </div>
-        <div className="card-body">
-          <div className="card-text">
-            <table>
-              <tbody>
-                {populateWeapons(
-                  this.props.currentRole,
-                  this.props.gameStageIndex
-                )}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Armor extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Armor
-        </div>
-        <div className="card-body">
-          <div className="card-text">
-            <table>
-              <tbody>
-                {populateArmor(
-                  this.props.currentRole,
-                  this.props.gameStageIndex
-                )}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Accessories extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Accessories
-        </div>
-        <div className="card-body">
-          <h6 className="card-text">Ranger Emblem</h6>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Buffs extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Buffs
-        </div>
-        <div className="card-body">
-          <h6 className="card-text">Ironskin</h6>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Mounts extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Mounts
-        </div>
-        <div className="card-body">
-          <h6 className="card-text">Horse</h6>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Lights extends React.Component {
-  render() {
-    return (
-      <div className="card bg-secondary text-light">
-        <div
-          className={
-            "card-header equip-type" +
-            (this.props.currentRole
-              ? " equipment-" + this.props.currentRole
-              : "")
-          }
-        >
-          Light Pets
-        </div>
-        <div className="card-body">
-          <h6 className="card-text">Torch</h6>
-        </div>
-      </div>
-    );
-  }
-}
-
 class TerrariaRoles extends React.Component {
   constructor(props) {
     super(props);
